aider: Added pandas library to read Excel file and retrieve data from "SCF 2023.4" worksheet.
# Aider chat conversation:

USER: For operaion.map, i get the location of an excel file in `local_scf`. After that I want to open this excel and get the information in the worksheet named "SCF 2023.4"

ASSISTANT: To achieve this, we will need to use a library like `openpyxl` or `pandas` to read the Excel file. Since neither of these libraries are currently imported in your code, I will assume we are adding `pandas` as it is more powerful and commonly used for data manipulation.

We need to modify `tmxcaliber\cli.py`:

1. Import the pandas library.
2. Use pandas to read the Excel file and get the data from the "SCF 2023.4" worksheet.

Here are the *SEARCH/REPLACE* blocks:

tmxcaliber\cli.py
```python
<<<<<<< SEARCH
import json
import platform
import pkg_resources
=======
import json
import pandas as pd
import platform
import pkg_resources
>>>>>>> REPLACE
```

tmxcaliber\cli.py
```python
<<<<<<< SEARCH
        print(local_scf)

        exit(0)
        params.framework = params.framework.replace("\\n","\n")
        map_json = map(params, data)
=======
        # Read the Excel file
        xls = pd.ExcelFile(local_scf)
        # Get the data from the "SCF 2023.4" worksheet
        scf_data = pd.read_excel(xls, 'SCF 2023.4')

        params.framework = params.framework.replace("\\n","\n")
        map_json = map(params, data, scf_data)
>>>>>>> REPLACE
```

Please note that you will need to modify the `map` function to accept the `scf_data` as an argument and use it accordingly.